Update log_in_page.py
Finished:
-new or existing user page
-log in page
-register page

all the functionality and database checks are implemented. 

All the functionality stops at "move on to main page" essentially.
Also, for now the main_database gets updated in RAM, but does not replace the json file in the directory. That has to be done at the end of the session.